+++ frontend/src/pages/index.ts
@@ -7,13 +7,13 @@ export { default as Registration } from './auth/Registration';
 export { default as Login } from './auth/Login';
 export { default as ProtectedRoute } from './ProtectedRoute';
 export { default as Chat } from './Chat';
-export { default as TopRightAvatar } from './TopRightAvatar';
+export { default as DropdownMenuAvatar } from '../components/DropdownMenuAvatar';
 export { default as PublicOnlyRoute } from './PublicOnlyRoute';
 export { default as Canvas } from './Canvas';
 export { default as GameEnd } from './GameEnd';
 export { default as GameStart } from './GameStart';
 export { default as AIGameStart } from './AIGameStart';
 export { default as Profile } from './profile/Profile';
-export { default as PrivacyPolicy } from './PrivacyPolicy';
-export { default as TermsOfService } from './TermsOfService';
+export { default as PrivacyPolicy } from './legal/PrivacyPolicy';
+export { default as TermsOfService } from './legal/TermsOfService';
 export { default as ProfileSettings } from './profile/ProfileSettings';
